duangb/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cloudinary==1.41.0
Django==5.1.1
django-ckeditor==6.7.1
django-debug-toolbar==4.4.6
django-js-asset==2.2.0
djangorestframework==3.15.2
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
pillow==10.4.0
psycopg2==2.9.9
pytz==2024.2
PyYAML==6.0.2
setuptools==75.1.0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
